Fix compilation errors.

+++ febft-messages/src/request_pre_processing/work_dividers/mod.rs
@@ -1,7 +1,6 @@
 use febft_communication::message::Header;
-use crate::messages::RequestMessage;
+use crate::messages::{ClientRqInfo, RequestMessage};
 use crate::request_pre_processing::{operation_key_raw, WorkPartitioner};
-use crate::timeouts::ClientRqInfo;
 
 pub struct WDRoundRobin;
 
@@ -9,12 +8,12 @@ impl<O> WorkPartitioner<O> for WDRoundRobin {
     fn get_worker_for(rq_info: &Header, message: &RequestMessage<O>, worker_count: usize) -> usize {
         let op_key = operation_key_raw(rq_info.from(), message.session_id());
 
-        (op_key % worker_count) as usize
+        (op_key % worker_count as u64) as usize
     }
 
     fn get_worker_for_processed(rq_info: &ClientRqInfo, worker_count: usize) -> usize {
         let op_key = operation_key_raw(rq_info.sender, rq_info.session);
 
-        (op_key % worker_count) as usize
+        (op_key % worker_count as u64) as usize
     }
 }
